requirements.txt
cmake
wheel
dlib-bin
Flask
Flask_Cors
matplotlib
numpy
opencv_contrib_python
opencv_python
scipy
